functions/authenticate_user/requirements.txt
PyJWT==2.8.0
pyotp==2.8.0
asyncpg==0.29.0
